requirements.txt
feedparser==6.0.10
beautifulsoup4==4.12.2
fastapi==0.104.0
uvicorn==0.23.2
pydantic==2.4.2
requests==2.31.0
python-dateutil==2.8.2
aiosqlite==0.19.0
keybert==0.7.0
spacy==3.7.2
jinja2==3.1.2 